% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphiR-package.R
\docType{package}
\name{delphiR}
\alias{delphiR}
\alias{delphiR-package}
\title{delphiR: R connector package for the Delphi project}
\description{
This is the R connector package for the Delphi project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/delphi-hub/delphiR}
  \item Report bugs at \url{https://github.com/delphi-hub/delphiR/issues}
}

}
\author{
\strong{Maintainer}: Ben Hermann \email{ben.hermann@googlemail.com}

}
\keyword{internal}
