% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphi_function.R
\name{features}
\alias{features}
\title{Retrieve the complete feature list supported by Delphi currently.}
\usage{
features(baseurl = httr::parse_url("https://delphi.cs.uni-paderborn.de/api"))
}
\arguments{
\item{baseurl}{The base URL of a Delphi service.}
}
\value{
A list containing a name and a description property for each feature.
}
\description{
Retrieve the complete feature list supported by Delphi currently.
}
\examples{
delphiR::features()
}
