% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphi_function.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve data for exactly one package}
\usage{
retrieve(
  id = "",
  baseurl = httr::parse_url("https://delphi.cs.uni-paderborn.de/api")
)
}
\arguments{
\item{id}{The maven identifier of the package requested}

\item{baseurl}{The base URL of a Delphi service.}
}
\value{
A list containing the request data.
}
\description{
Retrieve data for exactly one package
}
\examples{
delphiR::retrieve("log4j:log4j:1.2.7")
}
