% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphi_function.R
\name{search}
\alias{search}
\title{Search for matching packages.}
\usage{
search(
  query = "",
  baseurl = httr::parse_url("https://delphi.cs.uni-paderborn.de/api")
)
}
\arguments{
\item{query}{The query in Delphi query syntax.}

\item{baseurl}{The base URL of a Delphi service.}
}
\value{
A list containing the request data.
}
\description{
Search for matching packages.
}
\examples{
delphiR::search("[metrics.bytecode.ret]>0")
}
