% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphir.R
\name{set_delphi}
\alias{set_delphi}
\title{Set template files for delphi questionnaire}
\usage{
set_delphi(dir_name = "dir_name", jsPsych_version = "7.2.1")
}
\arguments{
\item{dir_name}{name of directory}

\item{jsPsych_version}{If you set a specific version number of jsPsych,
set_jsPsych prepare a file with that version of jsPsych.}
}
\description{
Set template files for delphi questionnaire
}
\examples{
# set_delphi("round1","7.2.1")
}
