% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphir.R
\name{set_round1}
\alias{set_round1}
\title{Set template files for delphi questionnaire}
\usage{
set_round1(
  inst,
  question,
  max_response = 10,
  min_response = 5,
  dir_name = "round1",
  jsPsych_version = "7.2.1"
)
}
\arguments{
\item{inst}{instruction}

\item{question}{question}

\item{max_response}{maximum number of response}

\item{min_response}{minimum number of response}

\item{dir_name}{name of directory}

\item{jsPsych_version}{If you set a specific version number of jsPsych,
set_jsPsych prepare a file with that version of jsPsych.}
}
\description{
Set template files for delphi questionnaire
}
\examples{
# set_round1()
}
