% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphi_function.R
\name{version}
\alias{version}
\title{Retrieve the version of a Delphi service.}
\usage{
version(baseurl = httr::parse_url("https://delphi.cs.uni-paderborn.de/api"))
}
\arguments{
\item{baseurl}{The base URL of a Delphi service.}
}
\value{
The version of the running service.
}
\description{
Retrieve the version of a Delphi service.
}
\examples{
delphiR::version()
}
